+++ tests/testbench-core/app-modules/test-module-two/src/Livewire/AnotherComponent.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace TestModuleTwo\Livewire;
+
+use Livewire\Component;
+
+class AnotherComponent extends Component
+{
+	public function render()
+	{
+		return '<div>Another Component</div>';
+	}
+}